se agrego el componente de footer
